refactor: refactor all pages
