+++ src/hooks/useNav.tsx
@@ -1,0 +1,9 @@
+import { useParams } from "react-router-dom";
+
+export const useNav = () => {
+  const { id } = useParams();
+
+  return {
+    id: Number(id),
+  };
+};
